Move reusable test classes into util
