+++ src/util/test/fakeRunWithLogContext.ts
@@ -16,11 +16,16 @@
 
 import { logger } from "@atomist/automation-client";
 import { RemoteRepoRef, RepoId } from "@atomist/automation-client/operations/common/RepoId";
-import { RunWithLogContext } from "../../../../src/common/delivery/goals/support/reportGoalError";
-import { ConsoleProgressLog } from "../../../../src/common/log/progressLogs";
-import { StatusForExecuteGoal, StatusState } from "../../../../src/typings/types";
-import { fakeContext } from "../../../software-delivery-machine/FakeContext";
+import { RunWithLogContext } from "../../common/delivery/goals/support/reportGoalError";
+import { ConsoleProgressLog } from "../../common/log/progressLogs";
+import { StatusForExecuteGoal, StatusState } from "../../typings/types";
+import { fakeContext } from "./fakeContext";
 
+/**
+ * Useful testing support
+ * @param {RemoteRepoRef} id
+ * @return {RunWithLogContext}
+ */
 export function fakeRunWithLogContext(id: RemoteRepoRef): RunWithLogContext {
     return {
         credentials: {token: "foobar"},
