+++ test/handlers/events/delivery/ExecuteGoalTest.ts
@@ -25,8 +25,8 @@ import { RunWithLogContext } from "../../../../src/common/delivery/goals/support
 import { createEphemeralProgressLog } from "../../../../src/common/log/EphemeralProgressLog";
 import { executeGoal } from "../../../../src/handlers/events/delivery/goals/executeGoal";
 import { SdmGoal } from "../../../../src/ingesters/sdmGoalIngester";
-import { SingleProjectLoader } from "../../../common/SingleProjectLoader";
-import { fakeContext } from "../../../software-delivery-machine/FakeContext";
+import { SingleProjectLoader } from "../../../../src/util/test/SingleProjectLoader";
+import { fakeContext } from "../../../../src/util/test/fakeContext";
 
 const helloWorldGoalExecutor = async (rwlc: RunWithLogContext) => {
     rwlc.progressLog.write("Hello world\n");
